Create Program.cs
Less time to intall latest Visual Studio.  Used DotnetFiddle to create the program moved the "FileDetails" class in Program to have the methods availabe here.
Please run this code with arguments passed in the debig section of the project. This will run perfectly as per the requirements if we pass the command line parameters to it.
I tested it with replcaing the "args" with hardcoded inputs, works as expected.